refactor layout using next route groups
