+++ __tests__/index.test.js
@@ -1,5 +1,5 @@
 import { render, screen } from "@testing-library/react";
-import Home from "../src/app/page";
+import Home from "../src/app/(main)/page";
 import "@testing-library/jest-dom";
 
 describe("Home", () => {
