+++ src/app/(main)/cookie-policy/page.tsx
