+++ src/app/(main)/layout.tsx
@@ -2,9 +2,9 @@ import Header from "@/components/header";
 import Banner from "@/components/banner";
 import CookiesBanner from "@/components/cookies";
 import GoogleAnalytics from "@/components/google-analytics";
-import { urlFor } from "./actions/client";
-import { getGlobalContent } from "./actions/actions";
-import "../styles/app.scss";
+import { urlFor } from "../actions/client";
+import { getGlobalContent } from "../actions/actions";
+import "../../styles/app.scss";
 import type { Metadata } from "next";
 import { cookies } from "next/headers";
