+++ src/app/(main)/planning-applications/[id]/feedback/page.tsx
@@ -2,9 +2,9 @@
 import { useEffect, useState } from "react";
 import Breadcrumbs from "@/components/breadcrumbs";
 import Instructions from "./instructions";
-import Questions from "../../../../components/questions";
-import { getLocalStorage } from "../../../lib/application";
-import { Data } from "../../../lib/type";
+import Questions from "../../../../../components/questions";
+import { getLocalStorage } from "../../../../lib/application";
+import { Data } from "../../../../lib/type";
 
 const Feedback = () => {
   const [application, setApplication] = useState<Data>();
