--- src/app/studio/[[...index]]/layout.tsx
@@ -1,13 +1,0 @@
-"use client";
-
-export default function SanityLayout({
-  children,
-}: {
-  children: React.ReactNode;
-}) {
-  return (
-    // <html lang="en">
-    <body>{children}</body>
-    // </html>
-  );
-}
